python/uv_setup/pyproject.toml
[project]
name = "uv-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
]

[tool.pytest.ini_options]
pythonpath = "src"
